+++ packages/server/src/auth/api/auth.ts
@@ -1,4 +1,3 @@
-import { render } from '@react-email/render'
 import { pick } from 'lodash'
 import { arg, intArg, mutationField, nonNull, stringArg } from 'nexus'
 import * as Yup from 'yup'
@@ -7,7 +6,7 @@ import { verifySignature } from '@/auth/model/signature'
 import { createAuthToken, createEmailToken } from '@/auth/model/token'
 import { Context } from '@/common/api'
 import { PIONEER_URL } from '@/common/config'
-import PioneerEmailTemplate from '@/common/email-templates/pioneer-email'
+import { renderPioneerEmail } from '@/common/email-templates/pioneer-email'
 import { createEmailProvider } from '@/common/utils/email'
 
 const emailProvider = createEmailProvider()
@@ -68,17 +67,15 @@ export const signup = mutationField('signup', {
       await emailProvider.sendEmail({
         to: args.email,
         subject: 'Confirm your email for Pioneer',
-        html: render(
-          PioneerEmailTemplate({
-            memberHandle: args.name,
-            summary: 'Confirm your email for Pioneer',
-            text: 'Please use the link below to confirm your email address for Pioneer notifications',
-            button: {
-              label: 'Confirm email',
-              href: verificationUrl,
-            },
-          })
-        ),
+        html: renderPioneerEmail({
+          memberHandle: args.name,
+          summary: 'Confirm your email for Pioneer',
+          text: 'Please use the link below to confirm your email address for Pioneer notifications',
+          button: {
+            label: 'Confirm email',
+            href: verificationUrl,
+          },
+        }),
       })
     }
