+++ packages/server/src/notifier/model/event/forum.ts
@@ -17,7 +17,6 @@ export const fromPostAddedEvent: NotifEventFromQNEvent<'PostAddedEvent'> = async
   const eventData = pick(postAddedEvent, 'inBlock', 'id')
 
   // TODO: add FORUM_POST_REPLY support
-  // TODO: somehow restrict number of notifs per event to avoid spamming
 
   return buildEvents(eventData, post.id, ({ generalEvent, entityEvent }) => [
     generalEvent('FORUM_POST_MENTION', mentionedMemberIds),
